pricepilot-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pricepilot"
version = "0.1.0"
description = "PricePilot - Google ADK agent for autonomous online grocery shopping"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0",
    "litellm",
    "playwright",
    "pydantic>=2.0",
    "fastapi",
    "uvicorn[standard]",
    "google-cloud-aiplatform",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["pricepilot"]
